+++ pystrich/code128/renderer.py
@@ -3,18 +3,11 @@
 from __future__ import annotations
 
 import logging
-from typing import TYPE_CHECKING
-
-from PIL import Image, ImageDraw, ImageFont
 
 from pystrich._vector_text import make_text_label
 from pystrich.bar_renderer import Bar1DRenderer
-from pystrich.fonts import get_font
-from pystrich.marks import BarLayout, TextLabel, iter_bar_marks
+from pystrich.marks import BarLayout, TextLabel
 from pystrich.types import BarcodeRenderOptions
-
-if TYPE_CHECKING:
-    from PIL.Image import Image as PILImage
 
 log = logging.getLogger("code128")
 
@@ -85,46 +78,3 @@ class Code128Renderer(Bar1DRenderer):
             quiet_bottom=label_border + fontsize + bottom_border,
             labels=labels,
         )
-
-    def get_pilimage(self, bar_width: int) -> PILImage:
-        """Return the barcode as a PIL object"""
-
-        show_label = self.options.get("show_label", True)
-        fontsize, _ = self._label_metrics(bar_width)
-        layout = self._bar_layout(bar_width)
-        log.debug("There are %d bars", len(layout.heights))
-
-        self.image_width = (
-            layout.quiet_left + len(layout.heights) * layout.bar_width + layout.quiet_right
-        )
-        self.image_height = self.options.get("height") or (self.image_width // 3)
-        bottom_border = self.options.get("bottom_border", 0)
-
-        img = Image.new("L", (self.image_width, self.image_height + bottom_border), 255)
-        draw = ImageDraw.Draw(img)
-
-        for mark in iter_bar_marks(
-            layout.heights,
-            layout.bar_width,
-            quiet_left=layout.quiet_left,
-            quiet_top=layout.quiet_top,
-        ):
-            draw.rectangle(
-                (mark.x, mark.y, mark.x + mark.width - 1, mark.y + mark.height - 1),
-                fill=0,
-            )
-
-        if show_label:
-            ttf_font = self.options.get("ttf_font")
-            font: ImageFont.ImageFont | ImageFont.FreeTypeFont
-            if ttf_font:
-                font = ImageFont.truetype(ttf_font, fontsize)
-            else:
-                font = get_font("courR", fontsize)
-            label_border = self.options.get("label_border", 0)
-            xtextwidth = font.getlength(self.text)
-            xtextpos = self.image_width / 2 - (xtextwidth / 2)
-            ytextpos = layout.quiet_top + max(layout.heights) + label_border
-            draw.text((xtextpos, ytextpos), self.text, font=font)
-
-        return img